+++ code/ui/container/registry-grid-container.tsx
@@ -6,24 +6,21 @@ import { useDictionary } from "@/hooks/useDictionary";
 import { Routes } from "@/io/config/routes";
 import { browserStorageManager } from "@/state/browser-storage-manager";
 import { Dictionary } from "@/types/dictionary";
-import { LifecycleStageMap, RegistryStatusMap } from "@/types/form";
+import { RegistryStatusMap } from "@/types/form";
 import { TableColumnOption } from "@/types/settings";
+import ViewAttachmentButton from "@/ui/graphic/table/action/view-attachment-button";
+import Button from "@/ui/interaction/button";
+import VirtualCard from "@/ui/interaction/card/virtual-card";
+import FilterMenu from "@/ui/interaction/menu/filter/filter-menu";
+import StatusComponent from "@/ui/text/status/status";
 import {
-  getInitialDateFromLifecycleStage,
   interpolate,
   parseWordsForLabels
 } from "@/utils/client-utils";
 import { Icon } from "@mui/material";
-import { ColumnFilter } from "@tanstack/react-table";
 import { useEffect } from "react";
-import LoadingSpinner from "../graphic/loader/spinner";
-import ViewAttachmentButton from "../graphic/table/action/view-attachment-button";
-import Accordion from "../interaction/accordion/accordion";
 import PopoverActionButton from "../interaction/action/popover/popover-button";
-import Button from "../interaction/button";
-import Card from "../interaction/card/card";
-import StatusComponent from "../text/status/status";
-import RegistryFilter from "./registry-filter";
+import DescriptionList from "../text/field/description-list";
 
 interface RegistryGridComponentProps {
   entityType: string;
@@ -40,11 +37,10 @@ export default function RegistryGridComponent(
   props: Readonly<RegistryGridComponentProps>
 ) {
   const dict: Dictionary = useDictionary();
-  const { isInitialLoading, parentRef, data, columns, selectedCount, totalCount, filters, virtualItems, rowVirtualizer,
-    resetFormSession, triggerRefresh, updateFilter, resetFilters } = useRegistryGrid(props.entityType, props.tableColumnOptions);
+  const { isInitialLoading, hasNoActiveFilters, parentRef, data, previewData, columns, currentItemIndex, selectedCount,
+    filters, virtualItems, rowVirtualizer, resetFormSession, triggerRefresh, updateFilter, resetFilters
+  } = useRegistryGrid(props.entityType, props.tableColumnOptions);
   const { navigateToDrawer } = useDrawerNavigation();
-  const hasNoActiveFilters: boolean = filters.filter(filter => filter?.id != "status")
-    .every((filter) => (filter?.value as string[])?.length == 0);
 
   useEffect(() => {
     // Trigger refresh when back navigation occurs
@@ -64,55 +60,15 @@ export default function RegistryGridComponent(
         <h1 className="py-1 md:py-4 text-2xl md:text-4xl font-bold">
           {parseWordsForLabels(props.entityType)}
         </h1>
-        {data.length > 0 && <PopoverActionButton
-          placement="bottom"
-          leftIcon="filter_list"
-          variant={hasNoActiveFilters ? "ghost" : "secondary"}
-          tooltipText={dict.action.filter}
-          size="icon"
-          aria-label={dict.action.filter}
-        >
-          <section className="flex justify-between ml-2 mr-4 items-center">
-            <h1>{dict.action.filter}</h1>
-            <Button
-              leftIcon="filter_list_off"
-              aria-label={dict.action.clearAllFilters}
-              iconSize="medium"
-              className="mt-1"
-              disabled={hasNoActiveFilters}
-              size="icon"
-              onClick={() => resetFilters()}
-              tooltipText={dict.action.clearAllFilters}
-              variant="destructive"
-            />
-          </section>
-          <section className="h-full max-h-100 overflow-y-auto">
-            {columns.map((column, index) => {
-              const fieldId: string = column.id;
-              const fieldTitle: string = column.header.toString();
-              const targetFilter: ColumnFilter = filters.find(filter => filter.id === fieldId);
-              const currentFilter: string[] = !targetFilter ? [] : (targetFilter.value as string[]);
-              return <Accordion
-                key={index}
-                id={fieldId}
-                title={fieldTitle}
-                isActive={currentFilter.length > 0}
-              >
-                <RegistryFilter
-                  type={props.entityType}
-                  field={fieldId}
-                  fieldType={column.dataType}
-                  lifecycleStage={LifecycleStageMap.OUTSTANDING}
-                  selectedDate={getInitialDateFromLifecycleStage(LifecycleStageMap.OUTSTANDING, false)}
-                  filters={filters}
-                  onSubmission={(selectedOptions: string[]) => {
-                    updateFilter(column.id.toString(), selectedOptions);
-                  }}
-                />
-              </Accordion>
-            })}
-          </section>
-        </PopoverActionButton>}
+        <FilterMenu
+          isInitialLoading={isInitialLoading}
+          hasNoActiveFilters={hasNoActiveFilters}
+          columns={columns}
+          entityType={props.entityType}
+          filters={filters}
+          resetFilters={resetFilters}
+          updateFilter={updateFilter}
+        />
       </div>
       {/** The virtualiser requires flex-1 height and overflow y in parent ref; it also requires an inner canvas container */}
       <section ref={parentRef} className="py-4 px-2 md:py-2.5 md:px-8 flex-1 overflow-y-auto min-h-0 w-full">
@@ -123,15 +79,14 @@ export default function RegistryGridComponent(
             position: "relative",
           }}
         >
-          {isInitialLoading && <LoadingSpinner isSmall={false} />}
-          {!isInitialLoading && data.length == 0 && <p className="p-2">{dict.message.noResultFound}</p>}
-          {data.length > 0 && virtualItems.map((virtualItem) => {
-            const isLoaderRow: boolean = virtualItem.index >= data.length;
+          {!isInitialLoading && previewData.length == 0 && <p className="p-2">{dict.message.noResultFound}</p>}
+          {!hasNoActiveFilters && previewData.length > 0 && virtualItems.map((virtualItem) => {
+            const isLoaderRow: boolean = virtualItem.index >= previewData.length;
             if (isLoaderRow) {
               return;
             }
-            const { id, date, event_id, status, ...displayFields } = data[virtualItem.index];
-            return <Card
+            const { id, date, event_id, status, ...displayFields } = previewData[virtualItem.index];
+            return <VirtualCard
               key={virtualItem.key}
               ref={rowVirtualizer.measureElement}
               data={displayFields}
@@ -171,28 +126,28 @@ export default function RegistryGridComponent(
                   navigateToDrawer(Routes.REGISTRY_TASK_COMPLETE, event_id);
                 }}
               />,
-              <Button
+              <PopoverActionButton
                 key={virtualItem.key + dict.action.view}
                 variant="ghost"
                 size="md"
                 iconSize="medium"
                 leftIcon="open_in_new"
                 label={parseWordsForLabels(dict.action.view)}
-                onClick={() => {
-                  browserStorageManager.clear();
-                  resetFormSession();
-                  navigateToDrawer(Routes.REGISTRY_TASK_VIEW, event_id);
-                }}
-              />]}
+              >
+                <section className="h-[80vh] overflow-y-auto">
+                  <DescriptionList
+                    data={data[virtualItem.index]}
+                  />
+                </section>
+              </PopoverActionButton>]}
             />
           })}
         </div>
       </section>
       <section className="flex justify-end">
         {data.length > 0 && <p className="text-sm pt-1 pr-4">{
-          interpolate(dict.message.numberOfRecords,
-            selectedCount != totalCount ? String(selectedCount) : String(totalCount))
-            .replace("{replacetotal}", String(totalCount))}</p>
+          interpolate(dict.message.numberOfRecords, String(currentItemIndex + 1))
+            .replace("{replacetotal}", String(selectedCount))}</p>
         }
       </section>
     </div>