+++ code/ui/global-container.tsx
@@ -5,12 +5,14 @@ import Konami from "react-konami-code";
 import { Provider } from "react-redux";
 
 import { reduxStore } from "@/app/store";
+import { usePullToRefresh } from "@/hooks/screen/usePullToRefresh";
 import { useBackgroundImageUrl } from "@/hooks/useBackgroundImageUrl";
 import { useContextMenu } from "@/hooks/useContextMenu";
+import { useOfflineWarning } from "@/hooks/useOfflineWarning";
 import { OptionalPage } from "@/io/config/optional-pages";
-import { usePathname } from "next/navigation";
 import { UISettings } from "@/types/settings";
 import Trex from "@/utils/trex";
+import { usePathname } from "next/navigation";
 import ContextMenu from "./interaction/context-menu/context-menu";
 import HeaderBar from "./interaction/header/headerbar";
 import { NavMenu } from "./navigation/navbar/nav-menu";
@@ -31,6 +33,8 @@ export default function GlobalContainer(props: Readonly<GlobalContainerProps>) {
   const backgroundImageUrl: string = useBackgroundImageUrl();
   const pathname = usePathname();
   const { contextMenuVisible, x: contextMenuX, y: contextMenuY, } = useContextMenu();
+  usePullToRefresh();
+  useOfflineWarning();
 
   const togglePopup = () => {
     setPopup(!popup);